refactor: use digital guide image in adapted toilets section
