+++ lib/features/digital_guide_view/tabs/adapted_toilets/data/repository/adapted_toilets_repository.dart
@@ -5,9 +5,7 @@ import "package:riverpod_annotation/riverpod_annotation.dart";
 import "../../../../../../api_base_rest/client/dio_client.dart";
 import "../../../../../../config/env.dart";
 import "../../../../data/models/digital_guide_response_extended.dart";
-import "../../../../data/repository/digital_guide_repository.dart";
 import "../models/adapted_toilet.dart";
-import "../models/adapted_toilet_not_full.dart";
 
 part "adapted_toilets_repository.g.dart";
 
@@ -32,19 +30,8 @@ Future<IMap<int, IList<AdaptedToilet>>> adaptedToiletsRepository(
           "${Env.digitalGuideUrl}/adapted_toilets/$adaptedTouiletID";
       final adaptedToiletNotFullResponse =
           await dio.get(adaptedToiletNotFullURL);
-      final adaptedToiletNotFull = AdaptedToiletNotFull.fromJson(
+      return AdaptedToilet.fromJson(
         adaptedToiletNotFullResponse.data as Map<String, dynamic>,
-      );
-      final imagesURLsFutures =
-          adaptedToiletNotFull.imagesIndices.map((imageID) async {
-        final imageURL = await ref.watch(GetImageUrlProvider(imageID).future);
-        return imageURL;
-      });
-      final imagesURLs =
-          (await Future.wait(imagesURLsFutures)).whereType<String>().toList();
-      return AdaptedToilet.create(
-        adaptedToiletNotFull: adaptedToiletNotFull,
-        imagesURLs: imagesURLs,
       );
     });
